Move view render logic into PageController class
